refactor : 각각의 대기 주문에 @Transactional이 적용되지 않는 문제 해결
